:recycle: [repositories] Extract function `git.clonerepository`
