+++ src/cutty/util/git.py
@@ -22,6 +22,44 @@ def initrepository(path: Path, *, head: Optional[str] = None) -> pygit2.Reposito
     return pygit2.init_repository(path, initial_head=head)
 
 
+def _fix_repository_head(repository: pygit2.Repository) -> pygit2.Reference:
+    """Work around a bug in libgit2 resulting in a bogus HEAD reference.
+
+    Cloning with a remote callback results in HEAD pointing to the user's
+    `init.defaultBranch` instead of the default branch of the cloned repository.
+    """
+    # https://github.com/libgit2/pygit2/issues/1073
+    head = repository.references["HEAD"]
+
+    with contextlib.suppress(KeyError):
+        return head.resolve()
+
+    for branch in ["main", "master"]:
+        ref = f"refs/heads/{branch}"
+        if ref in repository.references:
+            head.set_target(ref, message="repair broken HEAD after clone")
+            break
+
+    return head.resolve()
+
+
+def clonerepository(url: str, destination: Path) -> None:
+    """Clone a repository using a mirror configuration."""
+
+    def _createremote(
+        repository: pygit2.Repository, name: bytes, url: bytes
+    ) -> pygit2.Remote:
+        name_ = name.decode()
+        repository.config[f"remote.{name_}.mirror"] = True
+        return repository.remotes.create(name, url, "+refs/*:refs/*")
+
+    repository = pygit2.clone_repository(
+        url, str(destination), bare=True, remote=_createremote
+    )
+
+    _fix_repository_head(repository)
+
+
 def default_signature(repository: pygit2.Repository) -> pygit2.Signature:
     """Return the default signature."""
     with contextlib.suppress(KeyError):
